actions refactoring
