+++ users/views/author.py
@@ -1,17 +1,15 @@
-from django.shortcuts import get_object_or_404
-from django.utils import timezone
+from mixins import AuthorMixin
 from rest_framework import filters, viewsets
-from rest_framework.decorators import action
-from rest_framework.request import HttpRequest
-from rest_framework.response import Response
 
 from ..models.author import Author
 from ..serializers.author import AuthorSerializer
-from news.serializers.news import NewsSerializer
-from users.permission import AuthorOwnerOrReadOnly, IsModerator
+from users.permission import AuthorOwnerOrReadOnly
 
 
-class AuthorViewSet(viewsets.ModelViewSet):
+class AuthorViewSet(
+    viewsets.ModelViewSet,
+    AuthorMixin,
+):
     queryset = Author.objects.all()
     serializer_class = AuthorSerializer
     permission_classes = [AuthorOwnerOrReadOnly, ]
@@ -23,43 +21,3 @@ class AuthorViewSet(viewsets.ModelViewSet):
             email=self.request.user.email,
             user=self.request.user
         )
-
-    @action(
-        methods=['GET', ],
-        detail=True,
-        url_path='news',
-    )
-    def get_author_news(self, request: HttpRequest, pk: int):
-        author = get_object_or_404(Author, author_id=pk)
-        news = author.news.all()
-        serializer = NewsSerializer(news, many=True)
-        return Response(serializer.data)
-
-    @action(
-        detail=False,
-        methods=['GET', ],
-        url_path='author_stats',
-        permission_classes=[IsModerator, ]
-    )
-    def registration_counts(self, request: HttpRequest):
-        today = timezone.now().date()
-        week_start = today - timezone.timedelta(days=today.weekday())
-        month_start = today.replace(day=1)
-
-        today_count = self.queryset.filter(
-            date_joined__gte=today
-        ).count()
-        week_count = self.queryset.filter(
-            date_joined__gte=week_start
-        ).count()
-        month_count = self.queryset.filter(
-            date_joined__gte=month_start
-        ).count()
-
-        data = {
-            'today': today_count,
-            'this_week': week_count,
-            'this_month': month_count,
-        }
-
-        return Response(data)
