+++ users/views/user.py
@@ -1,16 +1,16 @@
+from mixins import UserMixin
 from rest_framework import viewsets
-from rest_framework.decorators import action
-from rest_framework.request import HttpRequest
-from rest_framework.response import Response
 
 from ..models.user import User
-from ..serializers.profile import ProfileSerializer
 from ..serializers.users import (UserCreateSerializer, UserListSerializer,
                                  UserUpdateSerializer)
 from users.permission import UserOwnerOrReadOnly
 
 
-class UserViewSet(viewsets.ModelViewSet):
+class UserViewSet(
+    viewsets.ModelViewSet,
+    UserMixin,
+):
     queryset = User.objects.all()
     permission_classes = [UserOwnerOrReadOnly, ]
 
@@ -20,26 +20,3 @@ class UserViewSet(viewsets.ModelViewSet):
         if self.action in ('update', 'partial_update'):
             return UserUpdateSerializer
         return UserListSerializer
-
-    @action(
-        methods=['GET'],
-        url_path='profile',
-        detail=False,
-        permission_classes=[UserOwnerOrReadOnly, ],
-    )
-    def profile(self, request: HttpRequest):
-        serializer = ProfileSerializer(request.user)
-        return Response(serializer.data)
-
-    @profile.mapping.patch
-    def update_my_model(self, request: HttpRequest):
-        serializer = ProfileSerializer(request.user, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        self.perform_update(serializer)
-        return Response(serializer.data)
-
-    @profile.mapping.delete
-    def profile_delete(self, request: HttpRequest):
-        serializer = ProfileSerializer(request.user)
-        serializer.destroy()
-        return Response({'detail': 'Success'}, status=204)
